Upmerge PR#83 to dev_ae_stdxl_ft (#86)
