+++ examples/stable-diffusion/training/README.md
@@ -115,7 +115,7 @@ image.save("cat-backpack.png")
 ```
 
 
-## Fine-Tuning
+## Fine-Tuning for SDXL
 
 The `train_text_to_image_sdxl.py` script shows how to implement the fine-tuning of Stable Diffusion models on Habana Gaudi.
 
@@ -126,15 +126,74 @@ Install the requirements:
 pip install -r requirements.txt
 ```
 
-### Example for SDXL
-We can launch the fine-tuning of SDXL model using:
+### Single-card Training
 
 ```bash
 python train_text_to_image_sdxl.py \
   --pretrained_model_name_or_path stabilityai/stable-diffusion-xl-base-1.0 \
   --pretrained_vae_model_name_or_path stabilityai/sdxl-vae \
   --dataset_name lambdalabs/pokemon-blip-captions \
-  --resolution 1024 \
+  --resolution 512 \
+  --crop_resolution 512 \
+  --center_crop \
+  --random_flip \
+  --proportion_empty_prompts=0.2 \
+  --train_batch_size 16 \
+  --max_train_steps 2500 \
+  --learning_rate 1e-05 \
+  --max_grad_norm 1 \
+  --lr_scheduler constant \
+  --lr_warmup_steps 0 \
+  --output_dir sdxl-pokemon-model \
+  --gaudi_config_name Habana/stable-diffusion \
+  --throughput_warmup_steps 3 \
+  --dataloader_num_workers 8 \
+  --bf16 \
+  --use_hpu_graphs_for_training \
+  --use_hpu_graphs_for_inference \
+  --validation_prompt="a robotic cat with wings" \
+  --validation_epochs 48 \
+  --checkpointing_steps 2500 \
+  --logging_step 10
+```
+
+
+### Multi-card Training
+```bash
+PT_HPU_RECIPE_CACHE_CONFIG=/tmp/stdxl_recipe_cache,True,1024  \
+python ../gaudi_spawn.py --world_size 8 --use_mpi train_text_to_image_sdxl.py \
+  --pretrained_model_name_or_path stabilityai/stable-diffusion-xl-base-1.0 \
+  --pretrained_vae_model_name_or_path stabilityai/sdxl-vae \
+  --dataset_name lambdalabs/pokemon-blip-captions \
+  --resolution 512 \
+  --crop_resolution 512 \
+  --center_crop \
+  --random_flip \
+  --proportion_empty_prompts=0.2 \
+  --train_batch_size 16 \
+  --max_train_steps 336 \
+  --learning_rate 1e-05 \
+  --max_grad_norm 1 \
+  --lr_scheduler constant \
+  --lr_warmup_steps 0 \
+  --output_dir sdxl-pokemon-model \
+  --gaudi_config_name Habana/stable-diffusion \
+  --throughput_warmup_steps 3 \
+  --dataloader_num_workers 8 \
+  --bf16 \
+  --use_hpu_graphs_for_training \
+  --use_hpu_graphs_for_inference \
+  --validation_prompt="a robotic cat with wings" \
+  --validation_epochs 48
+```
+
+### Single-card Training on Gaudi1
+```bash
+PT_HPU_MAX_COMPOUND_OP_SIZE=5 python train_text_to_image_sdxl.py \
+  --pretrained_model_name_or_path stabilityai/stable-diffusion-xl-base-1.0 \
+  --pretrained_vae_model_name_or_path stabilityai/sdxl-vae \
+  --dataset_name lambdalabs/pokemon-blip-captions \
+  --resolution 512 \
   --center_crop \
   --random_flip \
   --proportion_empty_prompts=0.2 \
@@ -148,64 +207,6 @@ python train_text_to_image_sdxl.py \
   --output_dir sdxl-pokemon-model \
   --gaudi_config_name Habana/stable-diffusion \
   --throughput_warmup_steps 3 \
-  --use_hpu_graphs \
   --bf16
 ```
 
-### Example for LoRA SDXL
-
-Low-Rank Adaption (LoRA) allows adapting a pretrained model by adding pairs of rank-decomposition matrices to
-existing weights and only training those newly added weights.
-
-We can launch the LoRA based fine-tuning of SDXL model using:
-
-```bash
-python train_text_to_image_sdxl.py \
-  --pretrained_model_name_or_path="stabilityai/stable-diffusion-xl-base-1.0" \
-  --pretrained_vae_model_name_or_path="madebyollin/sdxl-vae-fp16-fix" \
-  --dataset_name="lambdalabs/pokemon-blip-captions" \
-  --caption_column="text" \
-  --resolution=1024 --random_flip \
-  --train_batch_size=1 \
-  --num_train_epochs=2 --checkpointing_steps=500 \
-  --learning_rate=1e-04 --lr_scheduler="constant" --lr_warmup_steps=0 \
-  --seed=42 \
-  --output_dir="sd-pokemon-model-lora-sdxl" \
-  --finetuning_method="lora" \
-  --gaudi_config_name="Habana/stable-diffusion" \
-  --throughput_warmup_steps=3 \
-  --use_hpu_graphs \
-  --bf16
-```
-
-> [!NOTE]
-> SDXL's VAE is known to suffer from numerical instability issues. This is why we also expose a CLI argument namely `--pretrained_vae_model_name_or_path` that lets you specify the location of a better VAE (such as this one).
-
-#### LoRA SDXL Inference
-
-Once you have trained a LoRA weights as in the example above, inference can be done
-by using the `GaudiStableDiffusionXLPipeline`.
-
-```python
-import torch
-from optimum.habana.diffusers import (
-    GaudiStableDiffusionXLPipeline,
-    GaudiEulerDiscreteScheduler,
-)
-
-model_id = "stabilityai/stable-diffusion-xl-base-1.0"
-lora_model_id = "sd-pokemon-model-lora-sdxl"
-pipe = GaudiStableDiffusionXLPipeline.from_pretrained(
-  model_id,
-  scheduler=GaudiEulerDiscreteScheduler.from_pretrained(model_id, subfolder="scheduler"),
-  torch_dtype=torch.bfloat16,
-  use_habana=True,
-  use_hpu_graphs=True,
-  gaudi_config="Habana/stable-diffusion",
-)
-pipe.load_lora_weights(lora_model_id)
-
-prompt = "cute dragon creature"
-image = pipe(prompt).images[0]
-image.save("green-pokemon.png")
-```
