+++ examples/stable-diffusion/training/gpu/default_config_1x.yaml
