+++ examples/stable-diffusion/training/gpu/default_config_8x.yaml
