+++ examples/stable-diffusion/training/gpu/run_fp16.sh
