+++ examples/stable-diffusion/training/plot_loss_curve.py
