+++ examples/stable-diffusion/training/run.sh
