+++ examples/stable-diffusion/training/run_1x.sh
