+++ examples/stable-diffusion/training/train_text_to_image_sdxl.py
@@ -14,10 +14,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """
-Fine-tuning script for Stable Diffusion models for text2image with support for LoRA.
+Fine-tuning script for Stable Diffusion models for text2image.
 Adapted from the following sources:
-https://github.com/huggingface/diffusers/blob/v0.23.1/examples/text_to_image/train_text_to_image_sdxl.py
-https://github.com/huggingface/diffusers/blob/v0.23.1/examples/text_to_image/train_text_to_image_lora_sdxl.py
+https://github.com/huggingface/diffusers/blob/v0.25.1/examples/text_to_image/train_text_to_image_sdxl.py
 """
 
 import argparse
@@ -31,33 +30,32 @@ import random
 import shutil
 import time
 from pathlib import Path
-from typing import Dict
 
 import accelerate
 import datasets
 import diffusers
 import habana_frameworks.torch.core as htcore
+import habana_frameworks.torch as htorch
 import numpy as np
 import torch
 import torch.nn.functional as F
 import torch.utils.checkpoint
 import transformers
 from accelerate.logging import get_logger
-from accelerate.utils import DistributedDataParallelKwargs, ProjectConfiguration
+from accelerate.utils import ProjectConfiguration, DistributedDataParallelKwargs
+
 from datasets import load_dataset
 from diffusers import (
     AutoencoderKL,
+    DDPMScheduler,
     UNet2DConditionModel,
 )
-from diffusers.loaders import LoraLoaderMixin
 from diffusers.optimization import get_scheduler
 from diffusers.training_utils import EMAModel, compute_snr
 from diffusers.utils import check_min_version, is_wandb_available
 from diffusers.utils.torch_utils import is_compiled_module
 from huggingface_hub import create_repo, upload_folder
 from packaging import version
-from peft import LoraConfig
-from peft.utils import get_peft_model_state_dict
 from torchvision import transforms
 from torchvision.transforms.functional import crop
 from tqdm.auto import tqdm
@@ -66,8 +64,9 @@ from transformers import AutoTokenizer, PretrainedConfig
 from optimum.habana import GaudiConfig
 from optimum.habana.accelerate import GaudiAccelerator
 from optimum.habana.diffusers import GaudiEulerDiscreteScheduler, GaudiStableDiffusionXLPipeline
-from optimum.habana.utils import set_seed
-
+from optimum.habana.utils import set_seed, HabanaProfile
+from optimum.habana.accelerate.utils.dataclasses import GaudiDistributedType
+from optimum.habana.utils import to_gb_rounded
 
 if is_wandb_available():
     import wandb
@@ -82,7 +81,7 @@ DATASET_NAME_MAPPING = {
 }
 
 
-def save_model_card_sdxl(
+def save_model_card(
     repo_id: str,
     images=None,
     validation_prompt=None,
@@ -114,48 +113,6 @@ inference: true
 
 This pipeline was finetuned from **{base_model}** on the **{args.dataset_name}** dataset. Below are some example images generated with the finetuned pipeline using the following prompt: {validation_prompt}: \n
 {img_str}
-
-Special VAE used for training: {vae_path}.
-"""
-    with open(os.path.join(repo_folder, "README.md"), "w") as f:
-        f.write(yaml + model_card)
-
-
-def save_model_card_sdxl_lora(
-    repo_id: str,
-    images=None,
-    base_model=str,
-    dataset_name=str,
-    train_text_encoder=False,
-    repo_folder=None,
-    vae_path=None,
-):
-    img_str = ""
-    for i, image in enumerate(images):
-        image.save(os.path.join(repo_folder, f"image_{i}.png"))
-        img_str += f"![img_{i}](./image_{i}.png)\n"
-
-    yaml = f"""
----
-license: creativeml-openrail-m
-base_model: {base_model}
-dataset: {dataset_name}
-tags:
-- stable-diffusion-xl
-- stable-diffusion-xl-diffusers
-- text-to-image
-- diffusers
-- lora
-inference: true
----
-    """
-    model_card = f"""
-# LoRA text-to-image fine-tuning - {repo_id}
-
-These are LoRA adaption weights for {base_model}. The weights were fine-tuned on the {dataset_name} dataset. You can find some example images in the following. \n
-{img_str}
-
-LoRA for the text encoder was enabled: {train_text_encoder}.
 
 Special VAE used for training: {vae_path}.
 """
@@ -305,6 +262,15 @@ def parse_args(input_args=None):
         ),
     )
     parser.add_argument(
+        "--crop_resolution",
+        type=int,
+        default=1024,
+        help=(
+            "The resolution for crop input images, all the images in the train/validation dataset will be resized to this"
+            " resolution"
+        ),
+    )
+    parser.add_argument(
         "--center_crop",
         default=False,
         action="store_true",
@@ -319,11 +285,6 @@ def parse_args(input_args=None):
         help="whether to randomly flip images horizontally",
     )
     parser.add_argument(
-        "--train_text_encoder",
-        action="store_true",
-        help="Whether to train the text encoder. If set, the text encoder should be float32 precision.",
-    )
-    parser.add_argument(
         "--train_batch_size", type=int, default=16, help="Batch size (per device) for the training dataloader."
     )
     parser.add_argument("--num_train_epochs", type=int, default=100)
@@ -336,7 +297,7 @@ def parse_args(input_args=None):
     parser.add_argument(
         "--checkpointing_steps",
         type=int,
-        default=None,
+        default=500,
         help=(
             "Save a checkpoint of the training state every X updates. These checkpoints can be used both as final"
             " checkpoints in case they are better than the last checkpoint, and are also suitable for resuming"
@@ -519,19 +480,45 @@ def parse_args(input_args=None):
             " lazy mode."
         ),
     )
-    parser.add_argument("--use_hpu_graphs", action="store_true", help="Use HPU graphs on HPU.")
     parser.add_argument(
-        "--rank",
-        type=int,
-        default=4,
-        help=("The dimension of the LoRA update matrices."),
+        "--use_hpu_graphs_for_training",
+        action="store_true",
+        help="Use HPU graphs for training on HPU.")
+    parser.add_argument(
+        "--use_hpu_graphs_for_inference",
+        action="store_true",
+        help="Use HPU graphs for inference on HPU.")
+
+    parser.add_argument(
+        "--image_save_dir",
+        type=str,
+        default="./stable-diffusion-generated-images",
+        help="The directory where images will be saved.",
     )
     parser.add_argument(
-        "--finetuning_method",
+        "--output_type",
         type=str,
-        default="default",
-        choices=["default", "lora"],
-        help=("Set the method for fine-tuning." " Choices: ['default', 'lora']."),
+        choices=["pil", "np"],
+        default="pil",
+        help="Whether to return PIL images or Numpy arrays.",
+    )
+    parser.add_argument(
+        "--profiling_warmup_steps",
+        default=0,
+        type=int,
+        help="Number of steps to ignore for profiling.",
+    )
+    parser.add_argument(
+        "--profiling_steps",
+        default=0,
+        type=int,
+        help="Number of steps to capture for profiling.",
+    )
+    parser.add_argument(
+        "--logging_step",
+        default=1,
+        type=int,
+        help="Print the loss for every logging_step.",
     )
 
     if input_args is not None:
@@ -553,41 +540,8 @@ def parse_args(input_args=None):
     return args
 
 
-DATASET_NAME_MAPPING = {
-    "lambdalabs/pokemon-blip-captions": ("image", "text"),
-}
-
-
-def unet_attn_processors_state_dict(unet) -> Dict[str, torch.tensor]:
-    """
-    Returns:
-        a state dict containing just the attention processor parameters.
-    """
-    attn_processors = unet.attn_processors
-
-    attn_processors_state_dict = {}
-
-    for attn_processor_key, attn_processor in attn_processors.items():
-        for parameter_key, parameter in attn_processor.state_dict().items():
-            attn_processors_state_dict[f"{attn_processor_key}.{parameter_key}"] = parameter
-
-    return attn_processors_state_dict
-
-
-def tokenize_prompt(tokenizer, prompt):
-    text_inputs = tokenizer(
-        prompt,
-        padding="max_length",
-        max_length=tokenizer.model_max_length,
-        truncation=True,
-        return_tensors="pt",
-    )
-    text_input_ids = text_inputs.input_ids
-    return text_input_ids
-
-
 # Adapted from pipelines.StableDiffusionXLPipeline.encode_prompt
-def encode_prompt_sdxl(batch, text_encoders, tokenizers, proportion_empty_prompts, caption_column, is_train=True):
+def encode_prompt(batch, text_encoders, tokenizers, proportion_empty_prompts, caption_column, is_train=True):
     prompt_embeds_list = []
     prompt_batch = batch[caption_column]
 
@@ -623,51 +577,19 @@ def encode_prompt_sdxl(batch, text_encoders, tokenizers, proportion_empty_prompt
             bs_embed, seq_len, _ = prompt_embeds.shape
             prompt_embeds = prompt_embeds.view(bs_embed, seq_len, -1)
             prompt_embeds_list.append(prompt_embeds)
-
     prompt_embeds = torch.concat(prompt_embeds_list, dim=-1)
     pooled_prompt_embeds = pooled_prompt_embeds.view(bs_embed, -1)
-    return {"prompt_embeds": prompt_embeds.cpu(), "pooled_prompt_embeds": pooled_prompt_embeds.cpu()}
+    #map creates cache in cpu so need to change tensor to float32
+    return {"prompt_embeds": prompt_embeds.to(torch.float32), "pooled_prompt_embeds": pooled_prompt_embeds.to(torch.float32)}
 
 
-# Adapted from pipelines.StableDiffusionXLPipeline.encode_prompt
-def encode_prompt_sdxl_lora(text_encoders, tokenizers, prompt, text_input_ids_list=None):
-    prompt_embeds_list = []
-
-    for i, text_encoder in enumerate(text_encoders):
-        if tokenizers is not None:
-            tokenizer = tokenizers[i]
-            text_input_ids = tokenize_prompt(tokenizer, prompt)
-        else:
-            assert text_input_ids_list is not None
-            text_input_ids = text_input_ids_list[i]
-
-        prompt_embeds = text_encoder(
-            text_input_ids.to(text_encoder.device),
-            output_hidden_states=True,
-        )
-
-        # We are only ALWAYS interested in the pooled output of the final text encoder
-        pooled_prompt_embeds = prompt_embeds[0]
-        prompt_embeds = prompt_embeds.hidden_states[-2]
-        bs_embed, seq_len, _ = prompt_embeds.shape
-        prompt_embeds = prompt_embeds.view(bs_embed, seq_len, -1)
-        prompt_embeds_list.append(prompt_embeds)
-
-    prompt_embeds = torch.concat(prompt_embeds_list, dim=-1)
-    pooled_prompt_embeds = pooled_prompt_embeds.view(bs_embed, -1)
-    return prompt_embeds, pooled_prompt_embeds
-
-
-def compute_vae_encodings(batch, vae):
-    images = batch.pop("pixel_values")
-    pixel_values = torch.stack(list(images))
+def compute_vae_encodings(pixel_values, vae):
     pixel_values = pixel_values.to(memory_format=torch.contiguous_format).float()
     pixel_values = pixel_values.to(vae.device, dtype=vae.dtype)
-
     with torch.no_grad():
         model_input = vae.encode(pixel_values).latent_dist.sample()
     model_input = model_input * vae.config.scaling_factor
-    return {"model_input": model_input.cpu()}
+    return model_input
 
 
 def generate_timestep_weights(args, num_timesteps):
@@ -714,32 +636,16 @@ def generate_timestep_weights(args, num_timesteps):
 def main(args):
     logging_dir = Path(args.output_dir, args.logging_dir)
 
-    # Select appropriate pipeline version
-    sdxl_models = ["stable-diffusion-xl", "sdxl"]
-    sdxl = True if any(model in args.pretrained_model_name_or_path for model in sdxl_models) else False
+    accelerator_project_config = ProjectConfiguration(project_dir=args.output_dir, logging_dir=logging_dir)
 
-    if args.finetuning_method == "lora":
-        accelerator_project_config = ProjectConfiguration(project_dir=args.output_dir, logging_dir=logging_dir)
-        kwargs = DistributedDataParallelKwargs(find_unused_parameters=True)
-        gaudi_config = GaudiConfig.from_pretrained(args.gaudi_config_name)
-        accelerator = GaudiAccelerator(
-            gradient_accumulation_steps=args.gradient_accumulation_steps,
-            mixed_precision="bf16" if gaudi_config.use_torch_autocast or args.bf16 else "no",
-            log_with=args.report_to,
-            project_config=accelerator_project_config,
-            kwargs_handlers=[kwargs],
-            force_autocast=gaudi_config.use_torch_autocast or args.bf16,
-        )
-    else:
-        accelerator_project_config = ProjectConfiguration(project_dir=args.output_dir, logging_dir=logging_dir)
-        gaudi_config = GaudiConfig.from_pretrained(args.gaudi_config_name)
-        accelerator = GaudiAccelerator(
-            gradient_accumulation_steps=args.gradient_accumulation_steps,
-            mixed_precision="bf16" if gaudi_config.use_torch_autocast or args.bf16 else "no",
-            log_with=args.report_to,
-            project_config=accelerator_project_config,
-            force_autocast=gaudi_config.use_torch_autocast or args.bf16,
-        )
+    gaudi_config = GaudiConfig.from_pretrained(args.gaudi_config_name)
+    accelerator = GaudiAccelerator(
+        gradient_accumulation_steps=args.gradient_accumulation_steps,
+        mixed_precision="bf16" if gaudi_config.use_torch_autocast or args.bf16 else "no",
+        log_with=args.report_to,
+        project_config=accelerator_project_config,
+        force_autocast=gaudi_config.use_torch_autocast or args.bf16,
+    )
 
     # Make one log on every process with the configuration for debugging.
     logging.basicConfig(
@@ -794,7 +700,7 @@ def main(args):
     )
 
     # Load scheduler and models
-    noise_scheduler = GaudiEulerDiscreteScheduler.from_pretrained(
+    noise_scheduler = DDPMScheduler.from_pretrained(
         args.pretrained_model_name_or_path, subfolder="scheduler"
     )
 
@@ -805,6 +711,12 @@ def main(args):
     text_encoder_two = text_encoder_cls_two.from_pretrained(
         args.pretrained_model_name_or_path, subfolder="text_encoder_2", revision=args.revision, variant=args.variant
     ).to(accelerator.device)
+
+    # For mixed precision training we cast all non-trainable weigths to half-precision
+    # as these weights are only used for inference, keeping weights in full precision is not required.
+    weight_dtype = torch.float32
+    if gaudi_config.use_torch_autocast or args.bf16:
+        weight_dtype = torch.bfloat16
 
     vae_path = (
         args.pretrained_model_name_or_path
@@ -819,16 +731,21 @@ def main(args):
         variant=args.variant,
     )
     unet = UNet2DConditionModel.from_pretrained(
-        args.pretrained_model_name_or_path, subfolder="unet", revision=args.revision, variant=args.variant
+        args.pretrained_model_name_or_path,
+        subfolder="unet",
+        revision=args.revision,
+        variant=args.variant,
+        torch_dtype=weight_dtype
     )
 
     # Freeze vae and text encoders.
     vae.requires_grad_(False)
     text_encoder_one.requires_grad_(False)
     text_encoder_two.requires_grad_(False)
-
+    
     if args.gradient_checkpointing:
         unet.enable_gradient_checkpointing()
+
 
     if args.scale_lr:
         args.learning_rate = (
@@ -842,6 +759,11 @@ def main(args):
         optimizer_class = FusedAdamW
     else:
         optimizer_class = torch.optim.AdamW
+
+
+    if gaudi_config.use_fused_clip_norm:
+        from habana_frameworks.torch.hpex.normalization import FusedClipNorm
+        fused_clip_norm = FusedClipNorm(unet.parameters(), args.max_grad_norm)
 
     # Optimizer creation
     params_to_optimize = unet.parameters()
@@ -900,11 +822,19 @@ def main(args):
                 f"--caption_column' value '{args.caption_column}' needs to be one of: {', '.join(column_names)}"
             )
 
+    text_encoder_one = text_encoder_one.to(accelerator.device, dtype=weight_dtype)
+    text_encoder_two = text_encoder_two.to(accelerator.device, dtype=weight_dtype)
     # Preprocessing the datasets.
     train_resize = transforms.Resize(args.resolution, interpolation=transforms.InterpolationMode.BILINEAR)
     train_crop = transforms.CenterCrop(args.resolution) if args.center_crop else transforms.RandomCrop(args.resolution)
     train_flip = transforms.RandomHorizontalFlip(p=1.0)
     train_transforms = transforms.Compose([transforms.ToTensor(), transforms.Normalize([0.5], [0.5])])
+    
+    vae = vae.to(accelerator.device, dtype=weight_dtype)
+    # Let's first compute all the embeddings so that we can free up the text encoders
+    # from memory. We will pre-compute the VAE encodings too.
+    text_encoders = [text_encoder_one, text_encoder_two]
+    tokenizers = [tokenizer_one, tokenizer_two]
 
     def preprocess_train(examples):
         images = [image.convert("RGB") for image in examples[image_column]]
@@ -915,16 +845,21 @@ def main(args):
         for image in images:
             original_sizes.append((image.height, image.width))
             image = train_resize(image)
+            if args.crop_resolution < args.resolution:
+                if args.center_crop:
+                    y1 = max(0, int(round((image.height - args.resolution) / 2.0)))
+                    x1 = max(0, int(round((image.width - args.resolution) / 2.0)))
+                    image = train_crop(image)
+                else:
+                    y1, x1, h, w = train_crop.get_params(image, (args.resolution, args.resolution))
+                    image = crop(image, y1, x1, h, w)
+            else:
+                x1 = 0
+                y1 = 0
             if args.random_flip and random.random() < 0.5:
                 # flip
+                x1 = image.width - x1
                 image = train_flip(image)
-            if args.center_crop:
-                y1 = max(0, int(round((image.height - args.resolution) / 2.0)))
-                x1 = max(0, int(round((image.width - args.resolution) / 2.0)))
-                image = train_crop(image)
-            else:
-                y1, x1, h, w = train_crop.get_params(image, (args.resolution, args.resolution))
-                image = crop(image, y1, x1, h, w)
             crop_top_left = (y1, x1)
             crop_top_lefts.append(crop_top_left)
             image = train_transforms(image)
@@ -933,28 +868,6 @@ def main(args):
         examples["original_sizes"] = original_sizes
         examples["crop_top_lefts"] = crop_top_lefts
         examples["pixel_values"] = all_images
-        if args.finetuning_method == "lora":
-            # We need to tokenize input captions and transform the images.
-            def tokenize_captions(examples, is_train=True):
-                captions = []
-                for caption in examples[caption_column]:
-                    if isinstance(caption, str):
-                        captions.append(caption)
-                    elif isinstance(caption, (list, np.ndarray)):
-                        # take a random caption if there are multiple
-                        captions.append(random.choice(caption) if is_train else caption[0])
-                    else:
-                        raise ValueError(
-                            f"Caption column `{caption_column}` should contain either strings or lists of strings."
-                        )
-                tokens_one = tokenize_prompt(tokenizer_one, captions)
-                tokens_two = tokenize_prompt(tokenizer_two, captions)
-                return tokens_one, tokens_two
-
-            tokens_one, tokens_two = tokenize_captions(examples)
-            examples["input_ids_one"] = tokens_one
-            examples["input_ids_two"] = tokens_two
-
         return examples
 
     with accelerator.main_process_first():
@@ -963,65 +876,37 @@ def main(args):
         # Set the training transforms
         train_dataset = dataset["train"].with_transform(preprocess_train)
 
-    if args.finetuning_method == "lora":
+    compute_embeddings_fn = functools.partial(
+        encode_prompt,
+        text_encoders=text_encoders,
+        tokenizers=tokenizers,
+        proportion_empty_prompts=args.proportion_empty_prompts,
+        caption_column=args.caption_column,
+    )
 
-        def collate_fn(examples):
-            pixel_values = torch.stack([example["pixel_values"] for example in examples])
-            pixel_values = pixel_values.to(memory_format=torch.contiguous_format).bfloat16()  # float() to bfloat16()
-            original_sizes = [example["original_sizes"] for example in examples]
-            crop_top_lefts = [example["crop_top_lefts"] for example in examples]
-            input_ids_one = torch.stack([example["input_ids_one"] for example in examples])
-            input_ids_two = torch.stack([example["input_ids_two"] for example in examples])
-            return {
-                "pixel_values": pixel_values,
-                "input_ids_one": input_ids_one,
-                "input_ids_two": input_ids_two,
-                "original_sizes": original_sizes,
-                "crop_top_lefts": crop_top_lefts,
-            }
-    else:
-        # Let's first compute all the embeddings so that we can free up the text encoders
-        # from memory. We will pre-compute the VAE encodings too.
-        text_encoders = [text_encoder_one, text_encoder_two]
-        tokenizers = [tokenizer_one, tokenizer_two]
-        compute_embeddings_fn = functools.partial(
-            encode_prompt_sdxl,
-            text_encoders=text_encoders,
-            tokenizers=tokenizers,
-            proportion_empty_prompts=args.proportion_empty_prompts,
-            caption_column=args.caption_column,
-        )
-        compute_vae_encodings_fn = functools.partial(compute_vae_encodings, vae=vae)
-        with accelerator.main_process_first():
-            from datasets.fingerprint import Hasher
+    with accelerator.main_process_first():
+        from datasets.fingerprint import Hasher
 
-            # fingerprint used by the cache for the other processes to load the result
-            # details: https://github.com/huggingface/diffusers/pull/4038#discussion_r1266078401
-            new_fingerprint = Hasher.hash(args)
-            new_fingerprint_for_vae = Hasher.hash("vae")
+        # fingerprint used by the cache for the other processes to load the result
+        # details: https://github.com/huggingface/diffusers/pull/4038#discussion_r1266078401
+        new_fingerprint = Hasher.hash(args)
+        train_dataset = train_dataset.map(compute_embeddings_fn, batched=True,
+                                          new_fingerprint=new_fingerprint)
 
-            train_dataset = train_dataset.map(compute_embeddings_fn, batched=True, new_fingerprint=new_fingerprint)
-            train_dataset = train_dataset.map(
-                compute_vae_encodings_fn,
-                batched=True,
-                batch_size=args.train_batch_size * accelerator.num_processes * args.gradient_accumulation_steps,
-                new_fingerprint=new_fingerprint_for_vae,
-            )
+    def collate_fn(examples):
+        pixel_values = torch.stack([example["pixel_values"].clone().detach() for example in examples])
+        original_sizes = [example["original_sizes"] for example in examples]
+        crop_top_lefts = [example["crop_top_lefts"] for example in examples]
+        prompt_embeds = torch.stack([torch.tensor(example["prompt_embeds"]) for example in examples])
+        pooled_prompt_embeds = torch.stack([torch.tensor(example["pooled_prompt_embeds"]) for example in examples])
 
-        def collate_fn(examples):
-            model_input = torch.stack([torch.tensor(example["model_input"]) for example in examples])
-            original_sizes = [example["original_sizes"] for example in examples]
-            crop_top_lefts = [example["crop_top_lefts"] for example in examples]
-            prompt_embeds = torch.stack([torch.tensor(example["prompt_embeds"]) for example in examples])
-            pooled_prompt_embeds = torch.stack([torch.tensor(example["pooled_prompt_embeds"]) for example in examples])
-
-            return {
-                "model_input": model_input,
-                "prompt_embeds": prompt_embeds,
-                "pooled_prompt_embeds": pooled_prompt_embeds,
-                "original_sizes": original_sizes,
-                "crop_top_lefts": crop_top_lefts,
-            }
+        return {
+            "pixel_values": pixel_values,
+            "prompt_embeds": prompt_embeds,
+            "pooled_prompt_embeds": pooled_prompt_embeds,
+            "original_sizes": original_sizes,
+            "crop_top_lefts": crop_top_lefts,
+        }
 
     # DataLoaders creation:
     train_dataloader = torch.utils.data.DataLoader(
@@ -1032,162 +917,54 @@ def main(args):
         num_workers=args.dataloader_num_workers,
     )
 
-    if args.finetuning_method == "lora":
-        unet.requires_grad_(False)
+    # Set unet as trainable.
+    unet.train()
 
-        # For mixed precision training we cast all non-trainable weigths to half-precision
-        # as these weights are only used for inference, keeping weights in full precision is not required.
-
-        weight_dtype = torch.float32
-
-        if gaudi_config.use_torch_autocast or args.bf16:
-            weight_dtype = torch.bfloat16
-
-        # Move unet, vae and text_encoder to device and cast to weight_dtype
-        # The VAE is in float32 to avoid NaN losses.
-        unet.to(accelerator.device, dtype=weight_dtype)
-        vae.to(accelerator.device, dtype=weight_dtype)
-        text_encoder_one.to(accelerator.device, dtype=weight_dtype)
-        text_encoder_two.to(accelerator.device, dtype=weight_dtype)
-
-        # now we will add new LoRA weights to the attention layers
-        # Set correct lora layers
-        unet_lora_config = LoraConfig(
-            r=args.rank, init_lora_weights="gaussian", target_modules=["to_k", "to_q", "to_v", "to_out.0"]
+    del text_encoders, tokenizers
+    gc.collect()
+    # Create EMA for the unet.
+    if args.use_ema:
+        ema_unet = UNet2DConditionModel.from_pretrained(
+            args.pretrained_model_name_or_path, subfolder="unet", revision=args.revision, variant=args.variant
         )
-        unet.add_adapter(unet_lora_config)
+        ema_unet = EMAModel(ema_unet.parameters(), model_cls=UNet2DConditionModel, model_config=ema_unet.config)
 
-        # The text encoder comes from 🤗 transformers, so we cannot directly modify it.
-        # So, instead, we monkey-patch the forward calls of its attention-blocks.
-        if args.train_text_encoder:
-            # ensure that dtype is float32, even if rest of the model that isn't trained is loaded in fp16
-            text_lora_config = LoraConfig(
-                r=args.rank, init_lora_weights="gaussian", target_modules=["q_proj", "k_proj", "v_proj", "out_proj"]
-            )
-            text_encoder_one.add_adapter(text_lora_config)
-            text_encoder_two.add_adapter(text_lora_config)
+    # `accelerate` 0.16.0 will have better support for customized saving
+    if version.parse(accelerate.__version__) >= version.parse("0.16.0"):
+       # create custom saving & loading hooks so that `accelerator.save_state(...)` serializes in a nice format
+       def save_model_hook(models, weights, output_dir):
+           if accelerator.is_main_process:
+               if args.use_ema:
+                   ema_unet.save_pretrained(os.path.join(output_dir, "unet_ema"))
 
-        # create custom saving & loading hooks so that `accelerator.save_state(...)` serializes in a nice format
-        def save_model_hook(models, weights, output_dir):
-            if accelerator.is_main_process:
-                # there are only two options here. Either are just the unet attn processor layers
-                # or there are the unet and text encoder atten layers
-                unet_lora_layers_to_save = None
-                text_encoder_one_lora_layers_to_save = None
-                text_encoder_two_lora_layers_to_save = None
+               for i, model in enumerate(models):
+                   model.save_pretrained(os.path.join(output_dir, "unet"))
 
-                for model in models:
-                    if isinstance(model, type(accelerator.unwrap_model(unet))):
-                        unet_lora_layers_to_save = get_peft_model_state_dict(model)
-                    elif isinstance(model, type(accelerator.unwrap_model(text_encoder_one))):
-                        text_encoder_one_lora_layers_to_save = get_peft_model_state_dict(model)
-                    elif isinstance(model, type(accelerator.unwrap_model(text_encoder_two))):
-                        text_encoder_two_lora_layers_to_save = get_peft_model_state_dict(model)
-                    else:
-                        raise ValueError(f"unexpected save model: {model.__class__}")
+                   # make sure to pop weight so that corresponding model is not saved again
+                   weights.pop()
 
-                    # make sure to pop weight so that corresponding model is not saved again
-                    weights.pop()
+       def load_model_hook(models, input_dir):
+           if args.use_ema:
+               load_model = EMAModel.from_pretrained(os.path.join(input_dir, "unet_ema"), UNet2DConditionModel)
+               ema_unet.load_state_dict(load_model.state_dict())
+               ema_unet.to(accelerator.device)
+               del load_model
 
-                GaudiStableDiffusionXLPipeline.save_lora_weights(
-                    output_dir,
-                    unet_lora_layers=unet_lora_layers_to_save,
-                    text_encoder_lora_layers=text_encoder_one_lora_layers_to_save,
-                    text_encoder_2_lora_layers=text_encoder_two_lora_layers_to_save,
-                )
+           for i in range(len(models)):
+               # pop models so that they are not loaded again
+               model = models.pop()
 
-        def load_model_hook(models, input_dir):
-            unet_ = None
-            text_encoder_one_ = None
-            text_encoder_two_ = None
+               # load diffusers style into model
+               load_model = UNet2DConditionModel.from_pretrained(input_dir, subfolder="unet")
+               model.register_to_config(**load_model.config)
 
-            while len(models) > 0:
-                model = models.pop()
+               model.load_state_dict(load_model.state_dict())
+               del load_model
 
-                if isinstance(model, type(accelerator.unwrap_model(unet))):
-                    unet_ = model
-                elif isinstance(model, type(accelerator.unwrap_model(text_encoder_one))):
-                    text_encoder_one_ = model
-                elif isinstance(model, type(accelerator.unwrap_model(text_encoder_two))):
-                    text_encoder_two_ = model
-                else:
-                    raise ValueError(f"unexpected save model: {model.__class__}")
+       accelerator.register_save_state_pre_hook(save_model_hook)
+       accelerator.register_load_state_pre_hook(load_model_hook)
 
-            lora_state_dict, network_alphas = LoraLoaderMixin.lora_state_dict(input_dir)
-            LoraLoaderMixin.load_lora_into_unet(lora_state_dict, network_alphas=network_alphas, unet=unet_)
-
-            text_encoder_state_dict = {k: v for k, v in lora_state_dict.items() if "text_encoder." in k}
-            LoraLoaderMixin.load_lora_into_text_encoder(
-                text_encoder_state_dict, network_alphas=network_alphas, text_encoder=text_encoder_one_
-            )
-
-            text_encoder_2_state_dict = {k: v for k, v in lora_state_dict.items() if "text_encoder_2." in k}
-            LoraLoaderMixin.load_lora_into_text_encoder(
-                text_encoder_2_state_dict, network_alphas=network_alphas, text_encoder=text_encoder_two_
-            )
-
-        accelerator.register_save_state_pre_hook(save_model_hook)
-        accelerator.register_load_state_pre_hook(load_model_hook)
-    else:
-        # Set unet as trainable.
-        unet.train()
-
-        # For mixed precision training we cast all non-trainable weigths to half-precision
-        # as these weights are only used for inference, keeping weights in full precision is not required.
-
-        weight_dtype = torch.float32
-
-        if gaudi_config.use_torch_autocast or args.bf16:
-            weight_dtype = torch.bfloat16
-
-        # Move unet, vae and text_encoder to device and cast to weight_dtype
-        # The VAE is in float32 to avoid NaN losses.
-        vae.to(accelerator.device, dtype=weight_dtype)
-        text_encoder_one.to(accelerator.device, dtype=weight_dtype)
-        text_encoder_two.to(accelerator.device, dtype=weight_dtype)
-        del text_encoders, tokenizers
-        gc.collect()
-        # Create EMA for the unet.
-        if args.use_ema:
-            ema_unet = UNet2DConditionModel.from_pretrained(
-                args.pretrained_model_name_or_path, subfolder="unet", revision=args.revision, variant=args.variant
-            )
-            ema_unet = EMAModel(ema_unet.parameters(), model_cls=UNet2DConditionModel, model_config=ema_unet.config)
-
-        if version.parse(accelerate.__version__) >= version.parse("0.16.0"):
-            # create custom saving & loading hooks so that `accelerator.save_state(...)` serializes in a nice format
-            def save_model_hook(models, weights, output_dir):
-                if accelerator.is_main_process:
-                    if args.use_ema:
-                        ema_unet.save_pretrained(os.path.join(output_dir, "unet_ema"))
-
-                    for i, model in enumerate(models):
-                        model.save_pretrained(os.path.join(output_dir, "unet"))
-
-                        # make sure to pop weight so that corresponding model is not saved again
-                        weights.pop()
-
-            def load_model_hook(models, input_dir):
-                if args.use_ema:
-                    load_model = EMAModel.from_pretrained(os.path.join(input_dir, "unet_ema"), UNet2DConditionModel)
-                    ema_unet.load_state_dict(load_model.state_dict())
-                    ema_unet.to(accelerator.device)
-                    del load_model
-
-                for i in range(len(models)):
-                    # pop models so that they are not loaded again
-                    model = models.pop()
-
-                    # load diffusers style into model
-                    load_model = UNet2DConditionModel.from_pretrained(input_dir, subfolder="unet")
-                    model.register_to_config(**load_model.config)
-
-                    model.load_state_dict(load_model.state_dict())
-                    del load_model
-
-            accelerator.register_save_state_pre_hook(save_model_hook)
-            accelerator.register_load_state_pre_hook(load_model_hook)
-
+    
     # Scheduler and math around the number of training steps.
     overrode_max_train_steps = False
     num_update_steps_per_epoch = math.ceil(len(train_dataloader) / args.gradient_accumulation_steps)
@@ -1202,15 +979,12 @@ def main(args):
         num_training_steps=args.max_train_steps * args.gradient_accumulation_steps,
     )
 
+    unet = unet.to("hpu")
     # Prepare everything with our `accelerator`.
-    if args.finetuning_method == "lora" and args.train_text_encoder:
-        unet, text_encoder_one, text_encoder_two, optimizer, train_dataloader, lr_scheduler = accelerator.prepare(
-            unet, text_encoder_one, text_encoder_two, optimizer, train_dataloader, lr_scheduler
-        )
-    else:
-        unet, optimizer, train_dataloader, lr_scheduler = accelerator.prepare(
-            unet, optimizer, train_dataloader, lr_scheduler
-        )
+    unet, optimizer, train_dataloader, lr_scheduler = accelerator.prepare(
+        unet, optimizer, train_dataloader, lr_scheduler
+    )
+
 
     # We need to recalculate our total training steps as the size of the training dataloader may have changed.
     num_update_steps_per_epoch = math.ceil(len(train_dataloader) / args.gradient_accumulation_steps)
@@ -1218,17 +992,26 @@ def main(args):
         args.max_train_steps = args.num_train_epochs * num_update_steps_per_epoch
     # Afterwards we recalculate our number of training epochs
     args.num_train_epochs = math.ceil(args.max_train_steps / num_update_steps_per_epoch)
-
     # We need to initialize the trackers we use, and also store our configuration.
     # The trackers initializes automatically on the main process.
     if accelerator.is_main_process:
         accelerator.init_trackers("text2image-fine-tune-sdxl", config=vars(args))
 
-    def unwrap_model(model):
+    def unwrap_model(model, training=False):
         model = accelerator.unwrap_model(model)
         model = model._orig_mod if is_compiled_module(model) else model
-        return model
+        if not training:
+            return model
+        else:
+            if accelerator.distributed_type == GaudiDistributedType.MULTI_HPU:
+                kwargs = {}
+                kwargs["gradient_as_bucket_view"] = True
+                accelerator.ddp_handler = DistributedDataParallelKwargs(**kwargs)
+            if args.use_hpu_graphs_for_training:
+                htcore.hpu.ModuleCacher()(model=model, inplace=True)
 
+    unwrap_model(model=unet, training=True)
+    hb_profiler = HabanaProfile(warmup=args.profiling_warmup_steps, active=args.profiling_steps, record_shapes=False)
     # Train!
     total_batch_size = args.train_batch_size * accelerator.num_processes * args.gradient_accumulation_steps
 
@@ -1262,7 +1045,6 @@ def main(args):
         else:
             accelerator.print(f"Resuming from checkpoint {path}")
             accelerator.load_state(os.path.join(args.output_dir, path))
-
             global_step = int(path.split("-")[1])
 
             initial_global_step = global_step
@@ -1280,27 +1062,18 @@ def main(args):
     )
 
     t0 = None
+    t_start = time.perf_counter()
+    train_loss = torch.tensor(0, dtype=torch.float, device='hpu')
     for epoch in range(first_epoch, args.num_train_epochs):
-        if args.finetuning_method == "lora":
-            unet.train()
-            if args.train_text_encoder:
-                text_encoder_one.train()
-                text_encoder_two.train()
-        train_loss = 0.0
+        train_loss.zero_()
+        if hb_profiler:
+            hb_profiler.start()
         for step, batch in enumerate(train_dataloader):
-            if t0 is None:  # and global_step == args.throughput_warmup_steps:
+            if t0 is None or global_step == args.throughput_warmup_steps:
                 t0 = time.perf_counter()
-
             with accelerator.accumulate(unet):
-                if args.finetuning_method == "lora":
-                    # Convert images to latent space
-                    pixel_values = batch["pixel_values"].to(dtype=weight_dtype)
-                    model_input = vae.encode(pixel_values).latent_dist.sample()
-                    model_input = model_input * vae.config.scaling_factor
-                    model_input = model_input.to(weight_dtype)
-                else:
-                    model_input = batch["model_input"].to(dtype=weight_dtype).to(accelerator.device)
-
+                # Move compute_vae_encoding here to reflect the transformed image input
+                model_input = compute_vae_encodings(batch['pixel_values'], vae)
                 # Sample noise that we'll add to the latents
                 noise = torch.randn_like(model_input)
                 if args.noise_offset:
@@ -1309,34 +1082,27 @@ def main(args):
                     noise += args.noise_offset * torch.randn(
                         (model_input.shape[0], model_input.shape[1], 1, 1), device=rand_device
                     )
+                noise = noise.to(model_input.device)
 
                 bsz = model_input.shape[0]
 
-                # Sample a random timestep for each image
-                if args.finetuning_method == "lora":
+                if args.timestep_bias_strategy == "none":
+                    # Sample a random timestep for each image without bias.
                     timesteps = torch.randint(
                         0, noise_scheduler.config.num_train_timesteps, (bsz,), device=model_input.device
                     )
                     timesteps = timesteps.long()
                 else:
-                    if args.timestep_bias_strategy == "none":
-                        # Sample a random timestep for each image without bias.
-                        timesteps = torch.randint(
-                            0, noise_scheduler.config.num_train_timesteps, (bsz,), device=model_input.device
-                        )
-                        timesteps = timesteps.long()
-                    else:
-                        # Sample a random timestep for each image, potentially biased by the timestep weights.
-                        # Biasing the timestep weights allows us to spend less time training irrelevant timesteps.
-                        weights = generate_timestep_weights(args, noise_scheduler.config.num_train_timesteps).to(
-                            model_input.device
-                        )
-                        timesteps = torch.multinomial(weights, bsz, replacement=True).long()
+                    # Sample a random timestep for each image, potentially biased by the timestep weights.
+                    # Biasing the timestep weights allows us to spend less time training irrelevant timesteps.
+                    weights = generate_timestep_weights(args, noise_scheduler.config.num_train_timesteps).to(
+                        model_input.device
+                    )
+                    timesteps = torch.multinomial(weights, bsz, replacement=True).long()
 
                 # Add noise to the model input according to the noise magnitude at each timestep
                 # (this is the forward diffusion process)
                 noisy_model_input = noise_scheduler.add_noise(model_input, noise, timesteps)
-
                 # time ids
                 def compute_time_ids(original_size, crops_coords_top_left):
                     # Adapted from pipeline.StableDiffusionXLPipeline._get_add_time_ids
@@ -1349,21 +1115,10 @@ def main(args):
                 add_time_ids = torch.cat(
                     [compute_time_ids(s, c) for s, c in zip(batch["original_sizes"], batch["crop_top_lefts"])]
                 )
-
                 # Predict the noise residual
                 unet_added_conditions = {"time_ids": add_time_ids}
-
-                if args.finetuning_method == "lora":
-                    prompt_embeds, pooled_prompt_embeds = encode_prompt_sdxl_lora(
-                        text_encoders=[text_encoder_one, text_encoder_two],
-                        tokenizers=None,
-                        prompt=None,
-                        text_input_ids_list=[batch["input_ids_one"], batch["input_ids_two"]],
-                    )
-                else:
-                    prompt_embeds = batch["prompt_embeds"].to(accelerator.device)
-                    pooled_prompt_embeds = batch["pooled_prompt_embeds"].to(accelerator.device)
-
+                prompt_embeds = batch["prompt_embeds"].to(accelerator.device)
+                pooled_prompt_embeds = batch["pooled_prompt_embeds"].to(accelerator.device)
                 unet_added_conditions.update({"text_embeds": pooled_prompt_embeds})
 
                 model_pred = unet(
@@ -1383,7 +1138,7 @@ def main(args):
                     target = noise
                 elif noise_scheduler.config.prediction_type == "v_prediction":
                     target = noise_scheduler.get_velocity(model_input, noise, timesteps)
-                elif args.finetuning_method != "lora" and noise_scheduler.config.prediction_type == "sample":
+                elif noise_scheduler.config.prediction_type == "sample":
                     # We set the target to latents here, but the model_pred will return the noise sample prediction.
                     target = model_input
                     # We will have to subtract the noise residual from the prediction to get the target sample.
@@ -1411,31 +1166,31 @@ def main(args):
 
                 # Gather the losses across all processes for logging (if we use distributed training).
                 avg_loss = accelerator.gather(loss.repeat(args.train_batch_size)).mean()
-                train_loss += avg_loss.item() / args.gradient_accumulation_steps
-
+                train_loss += avg_loss / args.gradient_accumulation_steps
                 # Backpropagate
-                # TODO: check why this cause bufferoverflow issue
-                # with accelerator.autocast():
+                #TODO: check why this cause bufferoverflow issue
+                #with torch.autocast(device_type="hpu", dtype=weight_dtype, enabled=True):
                 accelerator.backward(loss)
                 htcore.mark_step()
 
                 if accelerator.sync_gradients:
-                    if args.finetuning_method == "lora":
-                        accelerator.clip_grad_norm_(params_to_optimize, args.max_grad_norm)
+                    params_to_clip = unet.parameters()
+                    if gaudi_config.use_fused_clip_norm:
+                        fused_clip_norm.clip_norm(params_to_clip)
                     else:
-                        params_to_clip = unet.parameters()
                         accelerator.clip_grad_norm_(params_to_clip, args.max_grad_norm)
+
                 optimizer.step()
                 lr_scheduler.step()
                 optimizer.zero_grad(set_to_none=True)
                 htcore.mark_step()
+                hb_profiler.step()
+
 
             # Checks if the accelerator has performed an optimization step behind the scenes
             if accelerator.sync_gradients:
                 progress_bar.update(1)
                 global_step += 1
-                accelerator.log({"train_loss": train_loss}, step=global_step)
-                train_loss = 0.0
 
                 if accelerator.is_main_process:
                     if args.checkpointing_steps is not None and global_step % args.checkpointing_steps == 0:
@@ -1463,73 +1218,65 @@ def main(args):
                         accelerator.save_state(save_path)
                         logger.info(f"Saved state to {save_path}")
 
-            logs = {"step_loss": loss.detach().item(), "lr": lr_scheduler.get_last_lr()[0]}
-            progress_bar.set_postfix(**logs)
+                if (global_step - 1) % args.logging_step == 0 or global_step == args.max_train_steps:
+                    train_loss_scalar = train_loss.item()
+                    accelerator.log({"train_loss": train_loss_scalar}, step=global_step)
+
+                    if args.gradient_accumulation_steps > 1:
+                        logs = {"step_loss": loss.item(), "lr": lr_scheduler.get_last_lr()[0], "mem_used": to_gb_rounded(htorch.hpu.memory_allocated())}
+                    else:
+                        logs = {"step_loss": train_loss_scalar, "lr": lr_scheduler.get_last_lr()[0], "mem_used": to_gb_rounded(htorch.hpu.memory_allocated())}
+                    progress_bar.set_postfix(**logs)
+                train_loss.zero_()
 
             if global_step >= args.max_train_steps:
                 break
 
+        hb_profiler.stop()
         if accelerator.is_main_process:
-            if args.validation_prompt is not None and epoch % args.validation_epochs == 0:
+            if args.validation_prompt is not None and (epoch+1) % args.validation_epochs == 0:
                 logger.info(
                     f"Running validation... \n Generating {args.num_validation_images} images with prompt:"
                     f" {args.validation_prompt}."
                 )
+                if args.use_ema:
+                    # Store the UNet parameters temporarily and load the EMA parameters to perform inference.
+                    ema_unet.store(unet.parameters())
+                    ema_unet.copy_to(unet.parameters())
 
-                if args.finetuning_method == "lora":
-                    # create pipeline
-                    pipeline = GaudiStableDiffusionXLPipeline.from_pretrained(
-                        args.pretrained_model_name_or_path,
-                        vae=vae,
-                        text_encoder=unwrap_model(text_encoder_one),
-                        text_encoder_2=unwrap_model(text_encoder_two),
-                        unet=accelerator.unwrap_model(unet),
-                        revision=args.revision,
-                        variant=args.variant,
-                        use_habana=True,
-                        use_hpu_graphs=args.use_hpu_graphs,
-                        gaudi_config=args.gaudi_config_name,
-                    )
-                    pipeline.scheduler = GaudiEulerDiscreteScheduler.from_config(pipeline.scheduler.config)
-                    pipeline = pipeline.to(accelerator.device)
-                    pipeline.set_progress_bar_config(disable=True)
-                else:
-                    if args.use_ema:
-                        # Store the UNet parameters temporarily and load the EMA parameters to perform inference.
-                        ema_unet.store(unet.parameters())
-                        ema_unet.copy_to(unet.parameters())
-
-                    # create pipeline
-                    vae = AutoencoderKL.from_pretrained(
-                        vae_path,
-                        subfolder="vae" if args.pretrained_vae_model_name_or_path is None else None,
-                        revision=args.revision,
-                        variant=args.variant,
-                    )
-                    pipeline = GaudiStableDiffusionXLPipeline.from_pretrained(
-                        args.pretrained_model_name_or_path,
-                        vae=vae,
-                        unet=unwrap_model(unet),
-                        revision=args.revision,
-                        variant=args.variant,
-                        use_habana=True,
-                        use_hpu_graphs=args.use_hpu_graphs,
-                        gaudi_config=args.gaudi_config_name,
-                    )
-                    if args.prediction_type is not None:
-                        scheduler_args = {"prediction_type": args.prediction_type}
-                        pipeline.scheduler = pipeline.scheduler.from_config(
-                            pipeline.scheduler.config, **scheduler_args
-                        )
-                    pipeline.scheduler = GaudiEulerDiscreteScheduler.from_config(pipeline.scheduler.config)
-                    pipeline = pipeline.to(accelerator.device)
-                    pipeline.set_progress_bar_config(disable=True)
+                # create pipeline
+                vae = AutoencoderKL.from_pretrained(
+                    vae_path,
+                    subfolder="vae" if args.pretrained_vae_model_name_or_path is None else None,
+                    revision=args.revision,
+                    variant=args.variant,
+                )
+                pipeline = GaudiStableDiffusionXLPipeline.from_pretrained(
+                    args.pretrained_model_name_or_path,
+                    vae=vae,
+                    unet=unwrap_model(unet),
+                    revision=args.revision,
+                    variant=args.variant,
+                    use_habana=True,
+                    use_hpu_graphs=args.use_hpu_graphs_for_inference,
+                    gaudi_config=args.gaudi_config_name,
+                )
+                if args.prediction_type is not None:
+                    scheduler_args = {"prediction_type": args.prediction_type}
+                    pipeline.scheduler = pipeline.scheduler.from_config(pipeline.scheduler.config, **scheduler_args)
+                pipeline.scheduler = GaudiEulerDiscreteScheduler.from_config(pipeline.scheduler.config)
+                pipeline = pipeline.to(accelerator.device)
+                pipeline.set_progress_bar_config(disable=True)
 
                 # run inference
+                generator = torch.Generator(device='cpu').manual_seed(args.seed) if args.seed else None
                 pipeline_args = {"prompt": args.validation_prompt}
 
-                with torch.autocast(device_type="hpu", dtype=weight_dtype, enabled=gaudi_config.use_torch_autocast):
-                    images = [pipeline(**pipeline_args).images[0] for _ in range(args.num_validation_images)]
+                with torch.autocast(device_type="hpu", dtype=torch.bfloat16, enabled=gaudi_config.use_torch_autocast):
+                    images = [
+                        pipeline(**pipeline_args, generator=generator, num_inference_steps=25).images[0]
+                        for _ in range(args.num_validation_images)
+                    ]
 
                 for tracker in accelerator.trackers:
                     if tracker.name == "tensorboard":
@@ -1548,12 +1295,14 @@ def main(args):
                 del pipeline
 
     duration = time.perf_counter() - t0
-    throughput = args.max_train_steps * total_batch_size / duration
+    ttt = time.perf_counter() - t_start
+    throughput = (args.max_train_steps - args.throughput_warmup_steps) * total_batch_size / duration
 
     accelerator.wait_for_everyone()
     if accelerator.is_main_process:
         logger.info(f"Throughput = {throughput} samples/s")
         logger.info(f"Train runtime = {duration} seconds")
+        logger.info(f"Total Train runtime = {ttt} seconds")
         metrics = {
             "train_samples_per_second": throughput,
             "train_runtime": duration,
@@ -1561,88 +1310,55 @@ def main(args):
         with open(f"{args.output_dir}/speed_metrics.json", mode="w") as file:
             json.dump(metrics, file)
 
-        if args.finetuning_method == "lora":
-            unet = accelerator.unwrap_model(unet)
+        unet = accelerator.unwrap_model(unet)
+        if args.use_ema:
+            ema_unet.copy_to(unet.parameters())
 
-            unet_lora_state_dict = get_peft_model_state_dict(unet)
-
-            if args.train_text_encoder:
-                text_encoder_one = accelerator.unwrap_model(text_encoder_one)
-                text_encoder_two = accelerator.unwrap_model(text_encoder_two)
-
-                text_encoder_lora_layers = get_peft_model_state_dict(text_encoder_one)
-                text_encoder_2_lora_layers = get_peft_model_state_dict(text_encoder_two)
-            else:
-                text_encoder_lora_layers = None
-                text_encoder_2_lora_layers = None
-
-            GaudiStableDiffusionXLPipeline.save_lora_weights(
-                save_directory=args.output_dir,
-                unet_lora_layers=unet_lora_state_dict,
-                text_encoder_lora_layers=text_encoder_lora_layers,
-                text_encoder_2_lora_layers=text_encoder_2_lora_layers,
-            )
-
-            del unet
-            del text_encoder_one
-            del text_encoder_two
-            del text_encoder_lora_layers
-            del text_encoder_2_lora_layers
-
-            pipeline = GaudiStableDiffusionXLPipeline.from_pretrained(
-                args.pretrained_model_name_or_path,
-                vae=vae,
-                revision=args.revision,
-                variant=args.variant,
-                torch_dtype=weight_dtype,
-                scheduler=noise_scheduler,
-                use_habana=True,
-                use_hpu_graphs=args.use_hpu_graphs,
-                gaudi_config=args.gaudi_config_name,
-            )
-            pipeline = pipeline.to(accelerator.device)
-
-            # load attention processors
-            pipeline.load_lora_weights(args.output_dir)
-        else:
-            unet = accelerator.unwrap_model(unet)
-            if args.use_ema:
-                ema_unet.copy_to(unet.parameters())
-
-            # Serialize pipeline.
-            vae = AutoencoderKL.from_pretrained(
-                vae_path,
-                subfolder="vae" if args.pretrained_vae_model_name_or_path is None else None,
-                revision=args.revision,
-                variant=args.variant,
-                torch_dtype=weight_dtype,
-            )
-            pipeline = GaudiStableDiffusionXLPipeline.from_pretrained(
-                args.pretrained_model_name_or_path,
-                unet=unet,
-                vae=vae,
-                revision=args.revision,
-                variant=args.variant,
-                torch_dtype=weight_dtype,
-                scheduler=noise_scheduler,
-                use_habana=True,
-                use_hpu_graphs=args.use_hpu_graphs,
-                gaudi_config=args.gaudi_config_name,
-            )
-            if args.prediction_type is not None:
-                scheduler_args = {"prediction_type": args.prediction_type}
-                pipeline.scheduler = pipeline.scheduler.from_config(pipeline.scheduler.config, **scheduler_args)
-            pipeline.save_pretrained(args.output_dir)
+        # Serialize pipeline.
+        vae = AutoencoderKL.from_pretrained(
+            vae_path,
+            subfolder="vae" if args.pretrained_vae_model_name_or_path is None else None,
+            revision=args.revision,
+            variant=args.variant,
+            torch_dtype=weight_dtype,
+        )
+        pipeline = GaudiStableDiffusionXLPipeline.from_pretrained(
+            args.pretrained_model_name_or_path,
+            unet=unet,
+            vae=vae,
+            revision=args.revision,
+            variant=args.variant,
+            torch_dtype=weight_dtype,
+            scheduler=noise_scheduler,
+            use_habana=True,
+            use_hpu_graphs=args.use_hpu_graphs_for_inference,
+            gaudi_config=args.gaudi_config_name,
+        )
+        if args.prediction_type is not None:
+            scheduler_args = {"prediction_type": args.prediction_type}
+            pipeline.scheduler = pipeline.scheduler.from_config(pipeline.scheduler.config, **scheduler_args)
+        pipeline.save_pretrained(args.output_dir)
 
         # run inference
         images = []
         if args.validation_prompt and args.num_validation_images > 0:
             pipeline = pipeline.to(accelerator.device)
+            generator = torch.Generator(device='cpu').manual_seed(args.seed) if args.seed else None
             with torch.autocast(device_type="hpu", dtype=weight_dtype, enabled=gaudi_config.use_torch_autocast):
                 images = [
-                    pipeline(args.validation_prompt, num_inference_steps=25).images[0]
+                    pipeline(args.validation_prompt, num_inference_steps=25, generator=generator).images[0]
                     for _ in range(args.num_validation_images)
                 ]
+            # Save images in the specified directory if not None and if they are in PIL format
+            if args.image_save_dir is not None:
+                if args.output_type == "pil":
+                    image_save_dir = Path(args.image_save_dir)
+                    image_save_dir.mkdir(parents=True, exist_ok=True)
+                    logger.info(f"Saving images in {image_save_dir.resolve()}...")
+                    for i, image in enumerate(images):
+                        image.save(image_save_dir / f"image_{epoch}_{i+1}.png")
+                else:
+                    logger.warning("--output_type should be equal to 'pil' to save images in --image_save_dir.")
 
             for tracker in accelerator.trackers:
                 if tracker.name == "tensorboard":
@@ -1659,27 +1375,15 @@ def main(args):
                     )
 
         if args.push_to_hub:
-            if sdxl:
-                if args.finetuning_method == "lora":
-                    save_model_card_sdxl_lora(
-                        repo_id=repo_id,
-                        images=images,
-                        base_model=args.pretrained_model_name_or_path,
-                        dataset_name=args.dataset_name,
-                        train_text_encoder=args.train_text_encoder,
-                        repo_folder=args.output_dir,
-                        vae_path=args.pretrained_vae_model_name_or_path,
-                    )
-                else:
-                    save_model_card_sdxl(
-                        repo_id=repo_id,
-                        images=images,
-                        validation_prompt=args.validation_prompt,
-                        base_model=args.pretrained_model_name_or_path,
-                        dataset_name=args.dataset_name,
-                        repo_folder=args.output_dir,
-                        vae_path=args.pretrained_vae_model_name_or_path,
-                    )
+            save_model_card(
+                repo_id=repo_id,
+                images=images,
+                validation_prompt=args.validation_prompt,
+                base_model=args.pretrained_model_name_or_path,
+                dataset_name=args.dataset_name,
+                repo_folder=args.output_dir,
+                vae_path=args.pretrained_vae_model_name_or_path,
+            )
             upload_folder(
                 repo_id=repo_id,
                 folder_path=args.output_dir,
